Basic Solar Model - CircularOrbit
Created a CircularOrbit class that, when initialized with a radius, length of year, and initial position can calculate its position for any given time.

Next step is to extend the class to work in three dimensions, accounting for the tilt relative to the invariable plane.

The datetime library has proven extremely useful as it allows both dates and lengths of time to interact seamlessly.